casting parser bug fixed.

+++ engine/engine.h
@@ -11,4 +11,4 @@
 
 void initialize_engine();
 void render_image(CMShader* shader, unsigned int texture_id, unsigned int vao, float x, float y, float width, float height, float rotation, int proj_width, int proj_height);
-void render_text(TTF_Font* font, CMShader* shader, std::wstring const& str, float x, float y, float _r, float _g, float _b, float _a, float rotation, int proj_width, int proj_height, int size);
+void render_text(TTF_Font* font, CMShader* shader, std::wstring const& str, float x, float y, int _r, int _g, int _b, int _a, float rotation, int proj_width, int proj_height, int size);
